Rearranges classes #3
