+++ src/main/java/com/lightweight/db/RandomDatabaseName.java
@@ -22,44 +22,65 @@
  * SOFTWARE.
  */
 
-package com.minlessika.lightweight.db;
+package com.lightweight.db;
 
 /**
- * Embedded HSQLDB DataSource.
- * @see <a href="http://h2database.com/html/features.html?highlight=HSQLDB%20Compatibility%20Mode&search=HSQLDB%20Compatibility%20Mode#firstFound">H2 - PostgreSQL Compatibility Mode</a>
+ * Random database name.
  * @since 0.1
- * @checkstyle AbbreviationAsWordInNameCheck (100 lines)
  */
-public final class EmbeddedHSQLDBDataSource extends DataSourceWrap {
+public final class RandomDatabaseName {
 
     /**
-     * Mode.
+     * Minimum of random interval.
      */
-    private static final String MODE = "HSQLDB";
+    private static final int DEFAULT_MIN = 0;
+
+    /**
+     * Maximum of random interval.
+     */
+    private static final int DEFAULT_MAX = 1_000_000;
+
+    /**
+     * Minimum of random interval.
+     */
+    private final int min;
+
+    /**
+     * Maximum of random interval.
+     */
+    private final int max;
 
     /**
      * Ctor.
+     * <p>Random interval.
      */
-    public EmbeddedHSQLDBDataSource() {
-        this(new RandomDatabaseName().value());
+    public RandomDatabaseName() {
+        this(
+            RandomDatabaseName.DEFAULT_MIN,
+            RandomDatabaseName.DEFAULT_MAX
+        );
     }
 
     /**
      * Ctor.
-     * @param dbname Database name
+     * @param min Minimum of random interval
+     * @param max Maximum of maximum interval
      */
-    public EmbeddedHSQLDBDataSource(final String dbname) {
-        this(dbname, EmbeddedDataSource.DEFAULT_MAX_POOL_SIZE);
+    public RandomDatabaseName(final int min, final int max) {
+        this.min = min;
+        this.max = max;
     }
 
     /**
-     * Ctor.
-     * @param dbname Database name
-     * @param maxpoolsize Max pool size
+     * Value.
+     * @return Name
      */
-    public EmbeddedHSQLDBDataSource(final String dbname, final int maxpoolsize) {
-        super(
-            new EmbeddedDataSource(dbname, EmbeddedHSQLDBDataSource.MODE, maxpoolsize)
+    public String value() {
+        final int randomnumber = this.min + (int) (Math.random() * ((this.max - this.min) + 1));
+        return String.format(
+            "test_db%s_%s",
+            Thread.currentThread().getId(),
+            randomnumber
         );
     }
 }
